reuse add_book_to_title to promote books without title
